Refactor limsdb queries and added project_info endpoint. Add status to Project to store open/closed status Add pages showing open/closed/all project status.
