+++ etc/column_mappings.yaml
@@ -34,6 +34,7 @@ column_def:
     matching_snps:            { data: 'genotype_validation.matching_snps',    title: 'Matching snps' ,                                                 visible: 'false'}
     median_coverage:          { data: 'median_coverage',                      title: 'Median coverage',          fmt: { type: 'float', min: {field: 'Coverage (X)', default: 30} } }
     mean_coverage:            { data: 'coverage.mean',                        title: 'Mean coverage',            fmt: { type: 'float', min: {field: 'Coverage (X)', default: 30} } }
+    closed_date:              { data: 'close_date',                           title: 'Close on',                 fmt: { type: 'date' }, visible: 'false' }
     coverage_5X:              { data: 'coverage',                             title: '% covered 5X',             fmt: { name: 'coverage_5X', type: 'percentage' }, visible: 'false' }
     coverage_15X:             { data: 'coverage',                             title: '% covered 15X',            fmt: { name: 'coverage_15X', type: 'percentage' } }
     evenness:                 { data: 'coverage.evenness',                    title: '% Evenness',               fmt: { type: 'ratio_percentage' } }
@@ -74,6 +75,7 @@ column_def:
     proc_status:              { data: 'aggregated.most_recent_proc.status',   title: 'Processing Status' }
     project_id:               { data: 'project_id',                           title: 'Project ID',               fmt: { link: '/projects/' }  }
     project_ids:              { data: 'project_ids',                          title: 'Projects',                 fmt: { link: '/projects/' } }
+    project_status:           { data: 'project_status',                       title: 'Status',                                                          visible: 'false'}
     plate_id:                 { data: 'plate_id',                             title: 'Plate' }
     properly_mapped_reads:    { data: 'properly_mapped_reads',                title: '# properly mapped reads',  fmt: { type: 'int' },    visible: 'false'}
     reviewed:                 { data: 'reviewed',                             title: 'Review status' }
@@ -155,7 +157,9 @@ project_status:
     - library_type
     - nb_invoiced
     - nb_genotyped
+    - project_status
     - open_date
+    - closed_date
     - started_date
     - finished_date
     - nb_removed
