+++ reporting_app/templates/reporting_base.html
@@ -37,8 +37,10 @@
                 <li>
                     <a href="#" class="dropdown=toggle" data-toggle="dropdown">Projects<b class="caret"></b></a>
                     <ul class="dropdown-menu">
-                        <li><a href="/projects">Projects</a></li>
-                        <li><a href="/project_status">Project Status</a></li>
+                        <li><a href="/projects">Project List</a></li>
+                        <li><a href="/project_status">Open Projects Status</a></li>
+                        <li><a href="/project_status/closed">Closed Projects Status</a></li>
+                        <li><a href="/project_status/all">All Projects Status</a></li>
                     </ul>
                 </li>
                 <li><a href="/charts">Charts</a></li>
