+++ rest_api/limsdb/queries/data_models.py
@@ -1,20 +1,59 @@
 import operator
 from datetime import datetime
 from collections import defaultdict
-from flask import request
-from egcg_core.clarity import sanitize_user_id
 from config import project_status as status_cfg
-from rest_api.common import retrieve_args
-from rest_api.limsdb import queries
 from rest_api.limsdb.queries import format_date
 
 
-class Sample:
+class SampleInfo:
     def __init__(self):
         self.sample_name = None
         self.project_name = None
         self.plate_name = None
         self.original_name = None
+        self.udfs = {}
+
+    def to_json(self):
+        to_return = {
+            'sample_id': self.sample_name,
+            'project_id': self.project_name,
+            'plate_id': self.plate_name,
+        }
+        to_return.update(self.udfs)
+        return to_return
+
+
+class ProjectInfo:
+
+    def __init__(self):
+        self.project_id = None
+        self.open_date = None
+        self.close_date = None
+        self.researcher_name = None
+        self.nb_quoted_samples = None
+
+    @property
+    def status(self):
+        if self.close_date:
+            return 'closed'
+        else:
+            return 'open'
+
+    def to_json(self):
+        ret = {
+            'project_id': self.project_id,
+            'open_date': format_date(self.open_date),
+            'close_date': format_date(self.close_date),
+            'project_status': self.status,
+            'researcher_name': self.researcher_name,
+            'nb_quoted_samples': self.nb_quoted_samples,
+        }
+        return ret
+
+
+class Sample(SampleInfo):
+    def __init__(self):
+        super().__init__()
         self._processes = set()
         self._status_and_date = None
         self._all_statuses_and_date = None
@@ -165,7 +204,7 @@ class Container:
     def __init__(self):
         self.samples = []
         self.project_id = None
-        self.name = None
+        self.container_name = None
 
     def samples_per_status(self):
         sample_per_status = defaultdict(list)
@@ -185,7 +224,7 @@ class Container:
 
     def to_json(self):
         ret = {
-            'plate_id': self.name,
+            'plate_id': self.container_name,
             'project_id': self.project_id,
             'nb_samples': len(self.samples),
             'library_type': self.library_types,
@@ -195,20 +234,23 @@ class Container:
         return ret
 
 
-class Project(Container):
+class Project(Container, ProjectInfo):
     def __init__(self):
         super().__init__()
         self.open_date = None
+        self.close_date = None
         self.researcher_name = None
         self.nb_quoted_samples = None
 
     def to_json(self):
         ret = {
-            'project_id': self.name,
+            'project_id': self.project_id,
             'nb_samples': len(self.samples),
             'library_type': self.library_types,
             'species': self.species,
             'open_date': format_date(self.open_date),
+            'close_date': format_date(self.close_date),
+            'project_status': self.status,
             'researcher_name': self.researcher_name,
             'nb_quoted_samples': self.nb_quoted_samples,
             'finished_date': format_date(self.finished_date),
@@ -238,98 +280,23 @@ class Project(Container):
         return None
 
 
-def _create_samples(session, match):
-    """This function queries the lims database for sample information and create Sample objects"""
-    all_samples = defaultdict(Sample)
-    project_id = match.get('project_id')
-    sample_id = match.get('sample_id')
-    sample_time_since = match.get('createddate')
-    if project_id or sample_id or sample_time_since:
-        only_open_project = False
-    else:
-        only_open_project = True
-    detailed = request.args.get('detailed') in ['true', 'True',  True]
-    if detailed:
-        list_process_complete = None
-        list_process_queued = None
-    else:
-        list_process_complete = list(status_cfg.step_completed_to_status) \
-                       + list(status_cfg.additional_step_completed) \
-                       + list(status_cfg.library_type_step_completed) \
-                       + status_cfg.started_steps
-        list_process_queued = status_cfg.step_queued_to_status
+class Run:
+    def __init__(self):
+        self.created_date = None
+        self.cst_date = None
+        self.udfs = {}
+        self.samples = set()
+        self.projects = set()
 
-    for result in queries.get_sample_info(session, project_id, sample_id, only_open_project=only_open_project,
-                                          time_since=sample_time_since, udfs=['Prep Workflow', 'Species']):
-        (pjct_name, sample_name, container, wellx, welly, udf_name, udf_value) = result
-        s = all_samples[sanitize_user_id(sample_name)]
-        s.sample_name = sanitize_user_id(sample_name)
-        s.project_name = pjct_name
-        s.plate_name = container
-        s.original_name = sample_name
-
-        if udf_name == 'Prep Workflow':
-            all_samples[sanitize_user_id(sample_name)].planned_library = udf_value
-        if udf_name == 'Species':
-            all_samples[sanitize_user_id(sample_name)].species = udf_value
-
-    for result in queries.get_samples_and_processes(session,  project_id, sample_id, only_open_project=only_open_project,
-                                                    workstatus='COMPLETE', list_process=list_process_complete,
-                                                    time_since=sample_time_since):
-        (pjct_name, sample_name, process_name, process_status, date_run, process_id) = result
-        all_samples[sanitize_user_id(sample_name)].add_completed_process(process_name, date_run, process_id)
-
-    for result in queries.get_sample_in_queues_or_progress(
-            session, project_id, sample_id, list_process=list_process_queued,
-            time_since=sample_time_since, only_open_project=only_open_project):
-        pjct_name, sample_name, process_name, queued_date, queue_id, process_id, process_date = result
-        if not process_id:
-            all_samples[sanitize_user_id(sample_name)].add_queue_location(process_name, queued_date, queue_id)
-        else:
-            all_samples[sanitize_user_id(sample_name)].add_inprogress(process_name, process_date, process_id)
-
-
-    return all_samples.values()
-
-
-def sample_status(session):
-    """This function queries the lims database for sample information and return json representation"""
-    kwargs = retrieve_args()
-    match = kwargs.get('match', {})
-
-    samples = _create_samples(session, match)
-    return [s.to_json() for s in samples]
-
-
-def sample_status_per_project(session):
-    """This function queries the lims database for sample information and aggregate at the project name level"""
-    kwargs = retrieve_args()
-    match = kwargs.get('match', {})
-    samples = _create_samples(session, match)
-    project_name = match.get('project_id')
-    all_projects = defaultdict(Project)
-    for project_info in queries.get_project_info(session, project_name, udfs=['Number of Quoted Samples']):
-        pjct_name, open_date, firstname, lastname, udf_name, nb_quoted_samples = project_info
-        all_projects[pjct_name].name = pjct_name
-        all_projects[pjct_name].open_date = open_date
-        all_projects[pjct_name].researcher_name = '%s %s' % (firstname, lastname)
-        all_projects[pjct_name].nb_quoted_samples = nb_quoted_samples
-
-    for sample in samples:
-        all_projects[sample.project_name].samples.append(sample)
-
-    return [p.to_json() for p in all_projects.values()]
-
-
-def sample_status_per_plate(session):
-    """This function queries the lims database for sample information and aggregate at the plate name level"""
-    kwargs = retrieve_args()
-    match = kwargs.get('match', {})
-    samples = _create_samples(session, match)
-    all_plates = defaultdict(Container)
-    for sample in samples:
-        all_plates[sample.plate_name].samples.append(sample)
-        all_plates[sample.plate_name].name = sample.plate_name
-        all_plates[sample.plate_name].project_id = sample.project_name
-
-    return [p.to_json() for p in all_plates.values()]
+    def to_json(self):
+        return {
+            'created_date': format_date(self.created_date),
+            'cst_date': format_date(self.cst_date),
+            'run_id': self.udfs['RunID'],
+            'run_status': self.udfs['Run Status'],
+            'sample_ids': sorted(list(self.samples)),
+            'project_ids': sorted(list(self.projects)),
+            'instrument_id': self.udfs['InstrumentID'],
+            'nb_reads': self.udfs['Read'],
+            'nb_cycles': self.udfs['Cycle']
+        }
